fix: vite hmr issues
