--- src/runtime/pages/stories-index.vue
@@ -1,17 +1,0 @@
-<template>
-  <StoriesLayout>
-    <div class="stories-index">
-      👋
-    </div>
-  </StoriesLayout>
-</template>
-
-<script setup lang="ts">
-import StoriesLayout from '../layouts/stories.vue'
-</script>
-
-<style scoped>
-.stories-index {
-  padding: 1rem;
-}
-</style>
