+++ src/runtime/pages/story-view.vue
@@ -1,36 +1,115 @@
 <template>
-  <StoriesLayout>
-    <div class="story-page">
-      <div
-        v-if="story"
-        class="story-content"
+  <div
+    class="stories-layout"
+    :data-bedtime-theme="theme"
+  >
+    <div class="stories-container">
+      <aside
+        ref="sidebarRef"
+        class="stories-sidebar"
       >
-        <component
-          :is="storyComponent"
-          v-if="storyComponent"
-        />
-      </div>
-      <div
-        v-else
-        class="error"
-      >
-        Story not found
-      </div>
+        <nav>
+          <ul>
+            <li
+              v-for="s in stories"
+              :key="s.kebabName"
+            >
+              <NuxtLink
+                :to="'/stories/' + s.kebabName"
+                class="stories-sidebar-link"
+                :class="{ active: currentStory === s.kebabName }"
+              >
+                {{ formatStoryName(s.pascalName) }}
+              </NuxtLink>
+            </li>
+          </ul>
+        </nav>
+      </aside>
+      <main class="stories-main">
+        <div class="story-page">
+          <div
+            v-if="story"
+            class="story-content"
+          >
+            <component
+              :is="storyComponent"
+              v-if="storyComponent"
+            />
+          </div>
+          <div
+            v-else-if="slug"
+            class="error"
+          >
+            Story not found
+          </div>
+        </div>
+      </main>
     </div>
-  </StoriesLayout>
+    <CommandPalette :items="storyItems" />
+  </div>
 </template>
 
 <script setup lang="ts">
-import { defineAsyncComponent } from 'vue'
-import { useStory } from '../composables/useStory'
-import StoriesLayout from '../layouts/stories.vue'
+import { computed, defineAsyncComponent, provide, ref, onMounted, onUpdated } from 'vue'
+import type { BedtimeStory } from '../../types/module'
 // @ts-expect-error resolved at runtime
-import { useHead, useRoute } from '#imports'
+import type { RouteLocationNormalized } from '#vue-router'
+// @ts-expect-error resolved at runtime
+import { useRoute, onBeforeRouteLeave, onBeforeRouteUpdate, useRuntimeConfig, useState, navigateTo, useHead } from '#imports'
+// @ts-expect-error virtual file
+import { stories as storyList } from '#build/stories.mjs'
 
 const route = useRoute()
-const { getStoryDetails } = useStory()
+const sidebarRef = ref<HTMLElement>()
+const stories = Object.values(storyList as Record<string, BedtimeStory>).sort((a, b) => a.pascalName.localeCompare(b.pascalName))
+const currentStory = computed(() => route.params.slug as string)
 
-const story = getStoryDetails(route.params.slug as string)
+const theme = useRuntimeConfig().public.bedtime?.viewer?.theme
+
+// This scroll position state stuff shouldn't be necessary
+// Maybe it's the fact that this isn't _really_ a Nuxt layout
+const lastScrollTop = useState<number>('bedtime-sidebar-scroll-top', () => 0)
+const saveScrollPosition = () => {
+  if (sidebarRef.value) {
+    lastScrollTop.value = sidebarRef.value.scrollTop
+  }
+}
+const restoreScrollPosition = () => {
+  if (sidebarRef.value) {
+    sidebarRef.value.scrollTop = lastScrollTop.value
+  }
+}
+onBeforeRouteLeave((_to: RouteLocationNormalized, _from: RouteLocationNormalized) => {
+  saveScrollPosition()
+})
+onBeforeRouteUpdate((_to: RouteLocationNormalized, _from: RouteLocationNormalized) => {
+  saveScrollPosition()
+})
+onMounted(() => {
+  restoreScrollPosition()
+})
+onUpdated(() => {
+  restoreScrollPosition()
+})
+
+function formatStoryName(name: string): string {
+  return name
+    .replace(/Story$/, '') // Remove 'Story' suffix
+    .trim()
+}
+
+const storyItems = computed(() => {
+  return stories.map(story => ({
+    id: story.kebabName,
+    label: story.pascalName,
+    action: () => { navigateTo(`/stories/${story.kebabName}`) },
+  }))
+})
+
+const slug = route.params.slug as string
+const story = storyList[slug]
+
+provide('story', story)
 
 const storyComponent = story?.component
   ? defineAsyncComponent(story.component)
@@ -42,6 +121,65 @@ useHead({
 </script>
 
 <style scoped>
+.stories-layout {
+  height: var(--stories-layout-height);
+  display: flex;
+  flex-direction: column;
+  overflow: hidden;
+}
+
+.stories-container {
+  display: flex;
+  flex: 1;
+  min-height: 0;
+}
+
+.stories-sidebar {
+  width: var(--stories-sidebar-width);
+  background-color: var(--stories-sidebar-bg-color);
+  border: var(--stories-sidebar-border);
+  padding: var(--stories-sidebar-padding);
+  overflow-y: auto;
+  flex-shrink: 0;
+}
+
+.stories-sidebar ul {
+  list-style: none;
+  padding: 0;
+  margin: 0;
+}
+
+.stories-sidebar li {
+  margin-top: 4px;
+  margin-bottom: 4px;
+}
+
+.stories-sidebar-link {
+  display: block;
+  padding: var(--stories-sidebar-link-padding);
+  color: var(--stories-sidebar-link-color);
+  text-decoration: none;
+  border-radius: var(--stories-sidebar-link-border-radius);
+  transition: var(--stories-sidebar-link-transition);
+  font-size: var(--stories-sidebar-link-font-size);
+}
+
+.stories-sidebar-link:hover {
+  background-color: var(--stories-sidebar-link-bg-hover);
+}
+
+.stories-sidebar-link.active {
+  background-color: var(--stories-sidebar-link-active-bg);
+  font-weight: var(--stories-sidebar-link-active-weight);
+}
+
+.stories-main {
+  flex: 1;
+  overflow-y: auto;
+  padding: var(--stories-main-padding);
+  min-height: 0;
+}
+
 .story-page {
   background-color: var(--stories-main-bg-color);
 }
